Move detekt-api/src/testFixtures to dev.detekt.api.test
