+++ detekt-api/src/testFixtures/kotlin/dev/detekt/api/testfixtures/TestSetupContext.kt
@@ -1,4 +1,4 @@
-package io.gitlab.arturbosch.detekt.test
+package dev.detekt.api.testfixtures
 
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.SetupContext
